+++ apps/ledger-live-desktop/src/renderer/families/bitcoin/SendAmountFields.tsx
@@ -11,7 +11,6 @@ import SelectFeeStrategy, { OnClickType } from "~/renderer/components/SelectFeeS
 import SendFeeMode from "~/renderer/components/SendFeeMode";
 import Text from "~/renderer/components/Text";
 import Tooltip from "~/renderer/components/Tooltip";
-import { context } from "~/renderer/drawers/Provider";
 import CoinControlModal from "./CoinControlModal";
 import { FeesField } from "./FeesField";
 import { BitcoinFamily } from "./types";
@@ -44,7 +43,6 @@ const Fields: Props = ({
 }) => {
   const bridge = getAccountBridge(account);
   const { t } = useTranslation();
-  const { state: drawerState, setDrawer } = React.useContext(context);
   const [coinControlOpened, setCoinControlOpened] = useState(false);
   const [isAdvanceMode, setAdvanceMode] = useState(!transaction.feesStrategy);
   const strategies = useFeesStrategy(account, transaction);
@@ -95,7 +93,6 @@ const Fields: Props = ({
           feesStrategy,
         }),
       );
-      if (drawerState.open) setDrawer(undefined);
     },
     // eslint-disable-next-line react-hooks/exhaustive-deps
     [updateTransaction, bridge],