+++ libs/ledger-live-common/src/featureFlags/defaultFeatures.ts
@@ -7,6 +7,7 @@ import {
 } from "@ledgerhq/types-live";
 import reduce from "lodash/reduce";
 import { formatToFirebaseFeatureId } from "./firebaseFeatureFlags";
+import { BUY_SELL_UI_APP_ID } from "../wallet-api/constants";
 
 /**
  * Default disabled feature.
@@ -384,14 +385,14 @@ export const DEFAULT_FEATURES: Features = {
   buySellUi: {
     enabled: false,
     params: {
-      manifestId: "multibuy-v2", // Update to "buy-sell-ui" after rollout
+      manifestId: BUY_SELL_UI_APP_ID,
     },
   },
 
   buySellShortcut: {
     enabled: false,
   },
-
+  ptxCard: DEFAULT_FEATURE,
   ptxSwapLiveAppDemoZero: {
     enabled: false,
     params: {