+++ libs/ledger-live-common/src/manager/api.ts
@@ -211,18 +211,16 @@ const getCurrentOSU: (input: {
 }) => Promise<OsuFirmware> = makeLRUCache(
   async input => {
     const { data } = await network({
-      method: "POST",
+      method: "GET",
       url: URL.format({
         pathname: `${getEnv("MANAGER_API_BASE")}/get_osu_version`,
         query: {
           livecommonversion,
+          device_version: input.deviceId,
+          version_name: `${input.version}-osu`,
+          provider: input.provider,
         },
       }),
-      data: {
-        device_version: input.deviceId,
-        version_name: `${input.version}-osu`,
-        provider: input.provider,
-      },
     });
     return data;
   },
@@ -239,18 +237,16 @@ const getCurrentFirmware: (input: {
     }: {
       data: FinalFirmware;
     } = await network({
-      method: "POST",
+      method: "GET",
       url: URL.format({
         pathname: `${getEnv("MANAGER_API_BASE")}/get_firmware_version`,
         query: {
           livecommonversion,
+          device_version: input.deviceId,
+          version_name: input.version,
+          provider: input.provider,
         },
       }),
-      data: {
-        device_version: input.deviceId,
-        version_name: input.version,
-        provider: input.provider,
-      },
     }).catch(error => {
       const status = error?.status || error?.response?.status;
 
@@ -290,17 +286,15 @@ const getDeviceVersion: (targetId: string | number, provider: number) => Promise
       }: {
         data: DeviceVersion;
       } = await network({
-        method: "POST",
+        method: "GET",
         url: URL.format({
           pathname: `${getEnv("MANAGER_API_BASE")}/get_device_version`,
           query: {
             livecommonversion,
+            provider,
+            target_id: targetId,
           },
         }),
-        data: {
-          provider,
-          target_id: targetId,
-        },
       }).catch(error => {
         const status = error?.status || error?.response?.status;